+++ src/xform-cli.ts
@@ -3,27 +3,46 @@ import { readFile, writeFile } from 'fs';
 import { join } from 'path';
 import { Templater } from '@guscrawford.com/json-xform';
 import { CliApp, StaticOption } from '@guscrawford.com/cleye';
+import { Stream } from 'stream';
+import { DefaultCommand } from './commands';
+import { read, parseVars } from './util';
+const open = require('open');
 let app = new CliApp(process.argv);
-let command = app.command({
-    name:'default'
+let help = app.command({
+    name:'help'
 });
-command.option(<StaticOption>{
-        spinalCaseName:'out',
-        flag:'o'
-    }
-).argument({
-    name:'template'
-})
-//console.info(command);
-readFile(join(process.cwd(),(command.args.template as any).value),{encoding:'utf8'},(err, data)=>{
+let command = new DefaultCommand(app).registeredCommand;
+const parse = (data:string)=>JSON.stringify(
+    new Templater(
+        parseVars(
+            JSON.parse(data),
+            (command.options.var as any).value
+        )
+    ).parse(),
+    null,
+    "  "
+);
+const transform = (err:any, data:string)=>{
     if (err) throw err;
     try {
-        writeFile(
-            join(process.cwd(),(command.options.out as any).value||(command.args.template as any).value), 
-            JSON.stringify(new Templater(JSON.parse(data)).parse(), null, "  "),
-            (err:any)=>{if (err)throw err}
-        );
+        let parsedData = parse(data);
+        if (!(command.options.out as any).value)
+            process.stdout.write(Buffer.from(parsedData+"\n\r"));
+        else
+            writeFile(
+                join(process.cwd(),(command.options.out as any).value||(command.args.template as any).value), 
+                parsedData,
+                (err:any)=>{if (err)throw err}
+            );
     } catch (e) {
         throw e;
     }
-});
+}
+if (help.index !== -1) {
+    let helpPath = app.binRuntime.split(/\/|\\/);
+        open(join(`${helpPath.slice(0, helpPath.length-2).join('/')}/docs/html/HELP.html`));
+}
+else if ((command.args.template as any).value)
+    readFile(join(process.cwd(),(command.args.template as any).value),{encoding:'utf8'},transform);
+else if (!(command.args.template as any).value)
+    read(process.stdin, transform);